+++ tests/BotManTest.php
@@ -1291,4 +1291,154 @@ class BotManTest extends PHPUnit_Framework_TestCase
         $botman->listen();
         $this->assertFalse($called_one);
     }
+
+    /** @test */
+    public function it_can_skip_a_running_conversation()
+    {
+        $called = false;
+        $driver = m::mock(NullDriver::class)->makePartial();
+
+        $driver->shouldReceive('getMessages')
+            ->andReturn([new Message('Hi Julia', 'UX12345', 'general')]);
+
+        $driver->shouldReceive('reply')
+            ->once()
+            ->with('This is a test question', m::type(Message::class), []);
+
+        $botman = $this->getBot([
+            'token' => 'foo',
+            'event' => [
+                'user' => 'UX12345',
+                'channel' => 'general',
+                'text' => 'Hi Julia',
+            ],
+        ]);
+
+        $botman->setDriver($driver);
+
+        $botman->hears('Hi Julia', function ($bot) {
+            $bot->startConversation(new TestConversation());
+        });
+        $botman->listen();
+
+        /*
+         * Now that the first message is saved, fake a reply.
+         * This should get skipped!
+         */
+        $botman = $this->getBot([
+            'token' => 'foo',
+            'event' => [
+                'user' => 'UX12345',
+                'channel' => 'general',
+                'text' => 'skip_keyword',
+            ],
+        ]);
+
+        $botman->hears('skip_keyword', function($bot) use (&$called) {
+            $called = true;
+        });
+
+        $botman->listen();
+        
+        $this->assertTrue($called);
+
+        $cacheKey = 'conversation-'.sha1('UX12345').'-'.sha1('general');
+        $this->assertTrue($this->cache->has($cacheKey));
+
+        /*
+         * This should now get picked up the usual way.
+         */
+        $called = false;
+        $botman = $this->getBot([
+            'token' => 'foo',
+            'event' => [
+                'user' => 'UX12345',
+                'channel' => 'general',
+                'text' => 'repeat',
+            ],
+        ]);
+
+        $botman->hears('repeat', function($bot) use (&$called) {
+            $called = true;
+        });
+
+        $botman->listen();
+
+        $this->assertFalse($called);
+    }
+
+    /** @test */
+    public function it_can_stop_a_running_conversation()
+    {
+        $called = false;
+        $driver = m::mock(NullDriver::class)->makePartial();
+
+        $driver->shouldReceive('getMessages')
+            ->andReturn([new Message('Hi Julia', 'UX12345', 'general')]);
+
+        $driver->shouldReceive('reply')
+            ->once()
+            ->with('This is a test question', m::type(Message::class), []);
+
+        $botman = $this->getBot([
+            'token' => 'foo',
+            'event' => [
+                'user' => 'UX12345',
+                'channel' => 'general',
+                'text' => 'Hi Julia',
+            ],
+        ]);
+
+        $botman->setDriver($driver);
+
+        $botman->hears('Hi Julia', function ($bot) {
+            $bot->startConversation(new TestConversation());
+        });
+        $botman->listen();
+
+        /*
+         * Now that the first message is saved, fake a reply.
+         * This should get skipped!
+         */
+        $botman = $this->getBot([
+            'token' => 'foo',
+            'event' => [
+                'user' => 'UX12345',
+                'channel' => 'general',
+                'text' => 'stop_keyword',
+            ],
+        ]);
+
+        $botman->hears('stop_keyword', function($bot) use (&$called) {
+            $called = true;
+        });
+        $botman->listen();
+
+        $this->assertTrue($called);
+
+        // Conversation should be removed from cache
+        $cacheKey = 'conversation-'.sha1('UX12345').'-'.sha1('general');
+        $this->assertFalse($this->cache->has($cacheKey));
+
+        /*
+         * This should now get picked up the usual way.
+         */
+        $called = false;
+        $botman = $this->getBot([
+            'token' => 'foo',
+            'event' => [
+                'user' => 'UX12345',
+                'channel' => 'general',
+                'text' => 'repeat',
+            ],
+        ]);
+
+        $botman->hears('repeat', function($bot) use (&$called) {
+            $called = true;
+        });
+
+        $botman->listen();
+
+        $this->assertTrue($called);
+    }
 }