+++ tests/Fixtures/TestConversation.php
@@ -3,6 +3,7 @@
 namespace Mpociot\BotMan\Tests\Fixtures;
 
 use Mpociot\BotMan\Answer;
+use Mpociot\BotMan\Message;
 use Mpociot\BotMan\Conversation;
 
 class TestConversation extends Conversation
@@ -20,6 +21,16 @@ class TestConversation extends Conversation
                 $this->repeat('This is a modified test question');
             }
         });
+    }
+
+    public function skipConversation(Message $message)
+    {
+        return $message->getMessage() === 'skip_keyword';
+    }
+
+    public function stopConversation(Message $message)
+    {
+        return $message->getMessage() === 'stop_keyword';
     }
 
     protected function _throwException($message)